requirements.txt
python-docx
langchain
chromadb
sentence_transformers
openai
weaviate-client
ragas==0.1.8